# docs/public/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_version_agents__agent_id__versions_upload_post": {
        "properties": {
          "data_access": {
            "default": "none",
            "title": "Data Access",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "registered_by": {
            "title": "Registered By",
            "type": "string"
          },
          "requested_hosts": {
            "default": "",
            "title": "Requested Hosts",
            "type": "string"
          },
          "tool_scope": {
            "default": "",
            "title": "Tool Scope",
            "type": "string"
          },
          "zip": {
            "format": "binary",
            "title": "Zip",
            "type": "string"
          }
        },
        "required": [
          "zip",
          "model_id",
          "registered_by"
        ],
        "title": "Body_upload_version_agents__agent_id__versions_upload_post",
        "type": "object"
      },
...
      "CredentialRequest": {
        "properties": {
          "agent_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Run Id"
          },
          "dataset_version_id": {
            "title": "Dataset Version Id",
            "type": "string"
          },
          "decide_only": {
            "default": false,
            "title": "Decide Only",
            "type": "boolean"
          },
          "principal": {
            "title": "Principal",
            "type": "string"
          },
          "principal_kind": {
            "enum": [
              "human",
              "workload"
            ],
            "title": "Principal Kind",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "title": "Purpose",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "run_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Secret"
          },
          "task_credential": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Credential"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "principal",
          "principal_kind",
          "roles",
          "tenant_id",
          "dataset_version_id"
        ],
        "title": "CredentialRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterAgentVersion": {
        "properties": {
          "code_hash": {
            "minLength": 1,
            "title": "Code Hash",
            "type": "string"
          },
          "image_digest": {
            "default": "native:host-venv",
            "title": "Image Digest",
            "type": "string"
          },
          "model_id": {
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "registered_by": {
            "title": "Registered By",
            "type": "string"
          },
          "requested_hosts": {
            "items": {
              "type": "string"
            },
            "title": "Requested Hosts",
            "type": "array"
          },
          "source_path": {
            "minLength": 1,
            "title": "Source Path",
            "type": "string"
          },
          "tool_scope": {
            "items": {
              "type": "string"
            },
            "title": "Tool Scope",
            "type": "array"
          }
        },
        "required": [
          "code_hash",
          "source_path",
          "model_id",
          "registered_by"
        ],
        "title": "RegisterAgentVersion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Munitas control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/agents/{agent_id}/versions": {
      "post": {
        "description": "Register and seal a version in one call.\n\nSealed on creation, the same as a dataset version, and for the same\nreason there is no update path anywhere in this file: a version somebody\ncould edit after the fact proves nothing about what actually ran.",
        "operationId": "register_version_agents__agent_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentVersion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Register Version Agents  Agent Id  Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Version",
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/{agent_id}/versions/upload": {
      "post": {
        "description": "Upload a project as this agent's next version, sealed on arrival.\n\n`code_hash` is not a claim here, unlike the declare-only path: it is\ncomputed from the bytes actually received, the same discipline\n`ingest.py` already applies to dataset uploads via `dataset_source.\nchecksum`. `tool_scope` arrives as a comma-separated string because this\nis a plain multipart form, not JSON, parsed and sorted the same way\n`scripts/admin/register-agent-version.py` already sorts its own `--tool` flags before\nsending them.",
        "operationId": "upload_version_agents__agent_id__versions_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_version_agents__agent_id__versions_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Upload Version Agents  Agent Id  Versions Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Version",
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/{agent_id}/versions/{version_id}/code": {
      "get": {
        "description": "Stream an uploaded version's code archive back, for the worker only.\n\nGated by a shared secret rather than a directory identity, because the\ncaller is the worker process itself, not a person or a registered\nagent principal, so \"who in the tenant's directory is this\" does not\napply to it. This endpoint is never reachable from `agentnet`: the\nsandboxed containers a version's code eventually runs inside have no\nroute to this API at all except `POST /credentials`, a different\nendpoint entirely.",
        "operationId": "download_code_agents__agent_id__versions__version_id__code_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-worker-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Code",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/credentials": {
      "post": {
        "description": "Mint a storage credential for one dataset version, or refuse and say why.\n\nThe order of operations matters and is not negotiable: decide, record the\ndecision, then act on it. Recording after acting would lose the audit trail\nfor exactly the requests that fail partway.\n\nThree outcomes: 200 with the key when access is allowed and in effect,\n403 when it is refused, and 202 when it is allowed but the storage\npermissions could not be updated yet. A 202 is never a refusal: the\nactivator finishes the job and resumes a run that parked on it.",
        "operationId": "request_credential_credentials_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-worker-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Credential"
      }
    },
...
  }
}